requirements.txt
pandas~=2.1.1
psycopg2~=2.9.7
SQLAlchemy~=2.0.21
selenium~=4.13.0
python-dotenv~=1.0.0
lxml~=4.9.3
requests~=2.31.0